refactor(prompt): split prompt builders behind interfaces
Extract the two oversized prompt builders into focused collaborators, using
the Strategy + Registry idiom already established for token estimators and
report renderers.

Attacker: each of the 16 per-attack-surface skill blocks becomes an
AttackerSkillInterface strategy under Prompt/Skill/ (ControllerAttackerSkill,
ApiResourceAttackerSkill, LiveComponentAttackerSkill, ...), self-declaring its
ProjectFileType and emission priority(). AttackerSkillRegistry collects them via
the symfony_security_auditor.attacker_skill DI tag and emits, in priority order,
the blocks whose type appears in the chunk. Adding an attack surface is now one
new tagged class with no edit to the builder. AttackerPromptBuilder drops from
640 to ~360 lines.

Reviewer: the fixed system-prompt text moves to ReviewerPromptSections and the
two line-numbered user-message templates to ReviewerMessageRenderer (both under
Prompt/Reviewer/, behind interfaces); ReviewerPromptBuilder is now pure per-mode
composition.

Emitted prompts are byte-identical (verified against every skill block);
PROMPT_VERSION is unchanged (attacker 11, reviewer 1) so no cached responses are
invalidated. All classes are @internal, so the internal restructure is not a BC
break. New AttackerSkillRegistryTest pins ordering and per-type emission.
