+++ src/Audit/Infrastructure/Prompt/AttackerPromptBuilder.php
@@ -18,6 +18,7 @@ use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\ProjectFile;
 use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\ProjectFileType;
 use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Model\SymfonyMapping;
 use VinceAmstoutz\SymfonySecurityAuditor\Audit\Domain\Port\AttackerPromptBuilderInterface;
+use VinceAmstoutz\SymfonySecurityAuditor\Audit\Infrastructure\Prompt\Skill\AttackerSkillRegistry;
 
 /** @internal not part of the BC promise — see docs/versioning.md */
 final readonly class AttackerPromptBuilder implements AttackerPromptBuilderInterface
@@ -37,296 +38,8 @@ final readonly class AttackerPromptBuilder implements AttackerPromptBuilderInter
     public function __construct(
         private bool $useStructuredCollection = self::DEFAULT_STRUCTURED_COLLECTION,
         private bool $emitAllSkills = self::DEFAULT_EMIT_ALL_SKILLS,
+        private AttackerSkillRegistry $attackerSkillRegistry = new AttackerSkillRegistry(),
     ) {}
-
-    /**
-     * Skill-block emission order — by attack-surface priority, NOT alphabetical.
-     * The LLM weights earlier-in-context instructions more heavily (primacy), so
-     * higher-risk surfaces are listed first.
-     *
-     * @var list<ProjectFileType>
-     */
-    private const array SKILL_PRIORITY = [
-        ProjectFileType::CONTROLLER,
-        ProjectFileType::API_RESOURCE,
-        ProjectFileType::LIVE_COMPONENT,
-        ProjectFileType::AUTHENTICATOR,
-        ProjectFileType::VOTER,
-        ProjectFileType::WEBHOOK_CONSUMER,
-        ProjectFileType::MESSENGER_HANDLER,
-        ProjectFileType::EVENT_SUBSCRIBER,
-        ProjectFileType::NORMALIZER,
-        ProjectFileType::SCHEDULER,
-        ProjectFileType::FORM,
-        ProjectFileType::REPOSITORY,
-        ProjectFileType::ENTITY,
-        ProjectFileType::TEMPLATE,
-        ProjectFileType::CONFIG,
-        ProjectFileType::PHP,
-    ];
-
-    /**
-     * Expert skill blocks injected into the system prompt when files of the matching
-     * ProjectFile::type() appear in the chunk. Each block lists both attack patterns
-     * to hunt AND patterns explicitly NOT to flag, reducing false positives.
-     *
-     * @var array<string, string>
-     */
-    private const array SKILLS = [
-        ProjectFileType::CONTROLLER->value => <<<'SKILL'
-            <skills role="controller">
-            Hunt:
-            - Missing `denyAccessUnlessGranted()` / `#[IsGranted]` on state-changing or sensitive read actions.
-            - `ParamConverter` / `MapEntity` / `#[MapRequestPayload]` / `#[MapQueryString]` flows for IDOR: path id → entity fetch with no ownership check.
-            - Mass-assignment: form `->submit($request->request->all())` without `allow_extra_fields: false` or restricted setters.
-            - `#[MapRequestPayload]` / `#[MapQueryString]` DTO mapping over an entity-shaped class with public mutable properties (mass-assignment via Serializer).
-            - Redirect targets and `RedirectResponse` arguments against open-redirect via user-controlled URLs.
-            - File-upload handlers: missing MIME validation, predictable upload paths, no content-type enforcement.
-            - Authentication endpoints (login, password reset, 2FA, registration) without rate-limiter binding (`#[IsSignatureValid]` / `RateLimiterFactory::create()` / `framework.rate_limiter`).
-            - `Request::get()` / `getContent()` flowing into Doctrine raw queries, `exec`, `passthru`, `eval`, `unserialize`, `simplexml_load_string`.
-            - Live Components: `#[LiveAction]` / `#[LiveProp(writable: true)]` exposing privileged setters or unbounded properties to the browser.
-            Do NOT flag:
-            - Controllers inheriting `denyAccessUnlessGranted()` from a parent class or a `#[IsGranted]` attribute on the class itself.
-            - Routes restricted by `methods: ['POST']` plus a CSRF-protected form — the form covers the CSRF concern.
-            - `#[MapRequestPayload]` / `#[MapQueryString]` over a DTO with validation constraints (`#[Assert\…]`) AND no privileged setters.
-            </skills>
-            SKILL,
-        ProjectFileType::AUTHENTICATOR->value => <<<'SKILL'
-            <skills role="authenticator">
-            Hunt:
-            - `authenticate()` returning a `SelfValidatingPassport` for password-bearing flows (skips credential check → auth bypass).
-            - `supports()` returning `null` instead of `false` for unsupported requests — `null` means "supports", silently letting non-matching paths through.
-            - `onAuthenticationSuccess()` calling `RedirectResponse` with a user-controlled `_target_path` / `referer` without protocol+host allowlist.
-            - `AccessTokenHandler::getUserBadgeFrom($accessToken)` not verifying signature/audience/issuer of JWT/opaque tokens.
-            - Custom `LoginFormAuthenticator` storing the password in the session, log, or exception trace.
-            - Missing `CsrfTokenBadge` on form-based login authenticators where CSRF was previously enforced.
-            - OAuth/OIDC handlers omitting `state` parameter verification (CSRF on auth callback) or PKCE for public clients.
-            - `UserBadge::setUserLoader()` query running with a user-controlled `identifier` without normalization (`User WHERE email = :id` accepting wildcards/case-insensitive variants).
-            Do NOT flag:
-            - Authenticators that explicitly throw `AuthenticationException` on missing credentials.
-            - `RememberMeBadge` on long-lived auth — that is the documented opt-in pattern.
-            </skills>
-            SKILL,
-        ProjectFileType::WEBHOOK_CONSUMER->value => <<<'SKILL'
-            <skills role="webhook_consumer">
-            Hunt:
-            - `RequestParserInterface::parse()` / `RemoteEventConsumerInterface::consume()` consuming the payload without HMAC / signature verification (`hash_equals` with the configured secret).
-            - Signature compared with `===` / `==` (timing-attack vulnerable) instead of `hash_equals()`.
-            - Missing replay-attack defense: no nonce check, no timestamp check, no idempotency key — same payload can be replayed indefinitely.
-            - `#[AsRemoteEventConsumer]` handler trusting `$payload['user_id']` / `$payload['amount']` without re-validating against the authenticated source.
-            - JSON / XML parsers without bounded depth/size (DoS via large or deeply-nested payloads).
-            - Webhook routes mounted under a firewall that allows anonymous access AND lacks IP allowlist or mutual-TLS gating.
-            Do NOT flag:
-            - Webhook handlers calling `hash_equals($expected, $received)` against the framework's secret.
-            - `WebhookComponent::validate($request, $secret)` invocations — those use constant-time comparison internally.
-            </skills>
-            SKILL,
-        ProjectFileType::MESSENGER_HANDLER->value => <<<'SKILL'
-            <skills role="messenger_handler">
-            Hunt:
-            - `#[AsMessageHandler]` / `MessageHandlerInterface::__invoke()` calling `unserialize()` / `igbinary_unserialize()` on payload fields.
-            - Handlers invoking `Process` / `shell_exec` / SQL with values from `$message` without sanitization (queue-to-shell injection).
-            - Missing idempotency: handler with side effects (charge card, send email, mutate balance) not deduping by message id (`AmqpStamp::getApplicationHeaders()['x-message-id']`).
-            - No replay protection: handler trusts `$message->createdAt` / `$message->userId` without verifying current state (stale message attack).
-            - Transport configured with `serializer: php` (PHP-native serialize on untrusted bus) — gadget-chain RCE.
-            - Handler swallowing exceptions silently → poisoned messages re-driven infinitely.
-            - Privileged action triggered solely by message presence with no authorization (`InvalidateUserCommand`, `PromoteToAdmin`, etc.).
-            Do NOT flag:
-            - Handlers using the default `JsonSerializer` transport — well-typed via Symfony Serializer.
-            - Handlers using `Symfony\Component\Messenger\Stamp\BusNameStamp` / `TransportNamesStamp` — those are routing, not security smells.
-            </skills>
-            SKILL,
-        ProjectFileType::EVENT_SUBSCRIBER->value => <<<'SKILL'
-            <skills role="event_subscriber">
-            Hunt:
-            - `KernelEvents::CONTROLLER` / `KernelEvents::REQUEST` subscribers mutating `$event->getRequest()->attributes` to inject privileged values (role, user id) before the controller runs.
-            - `SecurityEvents::AUTHENTICATION_SUCCESS` listeners auto-elevating roles based on payload fields without re-checking the source.
-            - `kernel.exception` listeners leaking stack traces, env vars, or internal hostnames in the response body.
-            - Subscribers calling `Process` / making HTTP requests with values from the event (SSRF via event payload).
-            - Listeners with side effects (DB write, mailer send) NOT wrapped in a Doctrine transaction or messenger envelope — request fails mid-way, state diverges.
-            - Doctrine `postLoad` / `postFlush` events writing user-derived fields back without escaping (stored XSS, log injection).
-            Do NOT flag:
-            - Subscribers using `$event->setResponse()` to short-circuit — that is the documented kernel.controller pattern.
-            - Listeners calling `LoggerInterface::info()` with structured arrays — not log injection unless raw `$_REQUEST` is interpolated.
-            </skills>
-            SKILL,
-        ProjectFileType::NORMALIZER->value => <<<'SKILL'
-            <skills role="normalizer">
-            Hunt:
-            - `denormalize()` building an Entity from request payload with `'allow_extra_attributes' => true` (mass-assignment) or without `'attributes' => [...]` allowlist.
-            - `denormalize()` calling private setters via reflection / `ObjectNormalizer` ignoring `#[Ignore]` on sensitive fields (`roles`, `passwordHash`, `isAdmin`).
-            - `normalize()` leaking sensitive fields by default (no `groups`, no `ignored_attributes`) — API leaks password hashes, tokens, internal ids.
-            - `supportsDenormalization()` returning `true` for `object` or untyped data — gadget-chain entry point.
-            - Custom denormalizer using `unserialize()` to decode a transport field.
-            - `getSupportedTypes()` returning `'*'` widely — denormalizer steals control from safer normalizers downstream.
-            Do NOT flag:
-            - `Symfony\Component\Serializer\Normalizer\PropertyNormalizer` without `setIgnoredAttributes()` when the model has only safe public properties.
-            - Normalizers operating purely on read-only DTOs with no setters.
-            </skills>
-            SKILL,
-        ProjectFileType::SCHEDULER->value => <<<'SKILL'
-            <skills role="scheduler">
-            Hunt:
-            - `#[AsSchedule]` / `ScheduleProviderInterface::getSchedule()` registering tasks that read user-input from DB and pass it unsanitized to `Process` / shell.
-            - Cron-like schedules invoking privileged operations (mass email, payout, account deletion) without an audit log or kill-switch.
-            - Tasks with no lock (`LockableTrait`, `LockFactory`) — overlapping runs cause duplicate billing or double notifications.
-            - Schedules running with `RunCommandMessage` whose `command` string is built from user input — RCE via the schedule.
-            - Recurring tasks fetching remote URLs (`HttpClient::request($urlFromDb)`) without a host allowlist (SSRF).
-            Do NOT flag:
-            - `RecurringMessage::every('1 hour', $message)` with a statically-typed message class and no user input.
-            - Schedules using `LockableTrait` with a project-unique lock name.
-            </skills>
-            SKILL,
-        ProjectFileType::VOTER->value => <<<'SKILL'
-            <skills role="voter">
-            Hunt:
-            - `supports($attribute, $subject)` mismatched with attributes consumers actually pass (typo → silent allow).
-            - `voteOnAttribute()` default branch returning `true` or implicit fallthrough — critical bypass.
-            - Over-broad role shortcuts (`ROLE_SUPER_ADMIN` always allowed) without auditable trail.
-            - Ownership checks with nullable / weak-typed comparisons (`==` on UUID strings, missing user object check).
-            - Voters consulting only `getRoles()` while the request requires per-resource attribute checks.
-            Do NOT flag:
-            - Voters that explicitly `return false` as the default — that is the secure-by-default pattern.
-            - Voters using `Security::isGranted('ROLE_USER')` after a positive ownership check.
-            </skills>
-            SKILL,
-        ProjectFileType::API_RESOURCE->value => <<<'SKILL'
-            <skills role="api_resource">
-            API Platform resources declare their entire HTTP surface in attributes — every operation is a routeless endpoint that never appears in a controller or access_control map. Hunt:
-            - Operations without `security:` — any `Get`/`GetCollection`/`Post`/`Patch`/`Put`/`Delete` in `operations: [...]`, as a standalone class attribute (e.g. `#[GetCollection]` without a wrapping `#[ApiResource]`), or as a resource-level default lacking a `security`/`securityPostDenormalize` expression is callable by anyone. Report as `broken_access_control`.
-            - Write operations relying only on `security:` — the expression runs BEFORE denormalization, against the object's OLD state. Ownership or role checks on writable data need `securityPostDenormalize:` (with `previous_object` where relevant) or an attacker updates other users' objects. Report as `broken_access_control` or `role_escalation`.
-            - Collection `GetCollection` without `security` or a Doctrine extension scoping the query to the current user — every record leaks. Report as `insecure_direct_object_reference`.
-            - `#[ApiFilter(SearchFilter::class, properties: [...])]` exposing sensitive or foreign-key properties (`user.id`, `email`, `token`) — filters become data-exfiltration oracles even when item access is denied. Report as `sensitive_data_exposure`.
-            - Serialization groups: `normalizationContext`/`denormalizationContext` groups placing privileged fields (`roles`, `isAdmin`, `passwordHash`, internal ids) on the read or write side. Write-side exposure is `over_permissive_serializer_group`.
-            - `paginationEnabled: false` (or client-controlled `paginationClientEnabled: true`) on large collections — unbounded result sets. Report as `missing_rate_limiting`.
-            - `Patch`/`Put` with `validationContext` groups that skip constraints enforced elsewhere.
-            - Custom state processors/providers that skip the voter/ownership checks the equivalent controller would perform.
-            Do NOT flag:
-            - Operations guarded at the firewall/access_control level when the pattern provably covers the resource route prefix.
-            - `security: "is_granted('PUBLIC_ACCESS')"` on intentionally public read-only catalogs.
-            - Filters on non-sensitive catalog fields (title, name, public slugs).
-            </skills>
-            SKILL,
-        ProjectFileType::LIVE_COMPONENT->value => <<<'SKILL'
-            <skills role="live_component">
-            Symfony UX Live Components are HTTP endpoints without routes: every `#[LiveAction]` method is invokable via POST by anyone who can reach the component, and every writable `#[LiveProp]` is bound from the request payload. Neither appears in a controller or the access_control map. Hunt:
-            - `#[LiveAction]` methods performing privileged work (delete, promote, pay, toggle flags) without `denyAccessUnlessGranted()` / an `#[IsGranted]` attribute on the method or class. The surrounding page's access control does NOT protect the component endpoint. Report as `broken_access_control`.
-            - `#[LiveProp(writable: true)]` on sensitive fields (ids of owned resources, prices, quantities, role-ish flags) — the client can set them to arbitrary values before an action runs. Report as `mass_assignment` or `price_manipulation`.
-            - Writable props flowing into queries, file paths, or redirects inside actions — standard injection sinks reached from a non-obvious source.
-            - `hydrateWith`/`dehydrateWith` custom hydration performing `unserialize` or object construction from client data. Report as `insecure_deserialization`.
-            - `#[LiveListener]` handlers trusting event payloads emitted client-side.
-            - Actions mutating entities loaded by client-supplied ids without an ownership check (IDOR through the component).
-            Do NOT flag:
-            - Read-only props (`#[LiveProp]` without `writable: true`) — the client cannot change them.
-            - Actions guarded by `#[IsGranted]` on the class or method, or an explicit `denyAccessUnlessGranted()` first statement.
-            - `#[AsTwigComponent]` (non-live) classes — they render server-side only and expose no endpoint.
-            </skills>
-            SKILL,
-        ProjectFileType::ENTITY->value => <<<'SKILL'
-            <skills role="entity">
-            Hunt:
-            - Lifecycle callbacks (`#[PrePersist]`, `#[PreUpdate]`) calling unsanitized methods, command execution, or external HTTP.
-            - Public setters on sensitive fields (`isAdmin`, `roles`, `passwordHash`) exposed via form normalization.
-            - Custom Doctrine types using `convertToPHPValue` with `unserialize` or `eval`-equivalent.
-            - DQL/QueryBuilder usage with string concatenation of request input — even inside the entity class.
-            - Boolean / enum coercion via Doctrine type juggling that lets attacker promote a role string.
-            - Serializer groups: `#[Groups([...])]` (or `@Groups({...})`) that places a privileged field (`roles`, `isAdmin`, `passwordHash`, `apiToken`, `internal*`) into a write-side group (e.g. `user:write`, `*:write`, `admin:*`) — denormalization will set it from request payload. Report as `over_permissive_serializer_group`.
-            - Read-side groups (`*:read`, `public`) leaking sensitive fields (`passwordHash`, `tokens`, internal ids) when the entity is serialized in API responses — also `over_permissive_serializer_group`.
-            Do NOT flag:
-            - Public setters on non-sensitive fields (titles, descriptions) where the form `mapped: false` or constraints validate input.
-            - `#[ORM\Column]` with `nullable: true` — nullability is a schema concern, not a security one.
-            - Read-only groups on safe fields (display name, public bio) — non-sensitive read groups are by design.
-            - Fields annotated `#[Ignore]` / `#[SerializedName]` redirecting to a public alias — those are explicit safe overrides.
-            </skills>
-            SKILL,
-        ProjectFileType::REPOSITORY->value => <<<'SKILL'
-            <skills role="repository">
-            Hunt:
-            - DQL/SQL injection via string concatenation in custom finder methods; absence of `setParameter()`.
-            - Dynamic `ORDER BY` / `LIMIT` fed from request input — Doctrine does NOT parameterize these.
-            - `NativeQuery` / `getConnection()->executeQuery()` with interpolated user input.
-            - Methods exposing arbitrary criteria (`findBy(['role' => $userInput])`) where caller forgot to constrain values.
-            - Subquery building that bypasses voters/filters applied at the controller layer.
-            Do NOT flag:
-            - QueryBuilder calls that use `setParameter()` for every user-controlled value — Doctrine binds those safely.
-            - `findOneBy([...])` / `findAll()` on non-sensitive entities; access control belongs to the voter, not the repository.
-            </skills>
-            SKILL,
-        ProjectFileType::FORM->value => <<<'SKILL'
-            <skills role="form">
-            Hunt:
-            - `'csrf_protection' => false` on forms processing state changes (acceptable only for stateless APIs with their own auth).
-            - `'allow_extra_fields' => true` enabling mass assignment of unmapped properties via setters.
-            - `EntityType` choice queries unscoped by ownership — attacker can select any other tenant's row.
-            - Data transformers that `unserialize` / `json_decode` raw input without strict mode.
-            - Missing `'constraints'` on free-form fields that flow into the database or shell.
-            Do NOT flag:
-            - Forms with the default CSRF token (Symfony enables it by default — only explicit `false` is the smell).
-            - Fields declared `mapped: false` and re-validated by a constraint — those don't reach the entity setter.
-            </skills>
-            SKILL,
-        ProjectFileType::TEMPLATE->value => <<<'SKILL'
-            <skills role="template">
-            Hunt:
-            - `|raw` filter applied to variables originating from user input or untrusted DB content.
-            - `autoescape` overridden to `false` or to a context that does not match the surrounding HTML/JS/URL context.
-            - `{{ include(user_input) }}` or `{% include %}` with dynamic template names — SSTI vector.
-            - `{% sandbox %}` / `{% apply %}` blocks lifting restrictions on user-supplied template fragments.
-            - Inline JavaScript context (`<script>var x = {{ value }};`) without `|json_encode` — XSS via Twig.
-            - URL attributes (`href`, `src`) built from user input without `|url_encode` and protocol whitelist (javascript:, data:).
-            - Twig Components (`<twig:Component …/>`) passing user input through `data-*` attributes without escaping (`<twig:UserCard name="{{ name|raw }}"/>`).
-            - Live Components emitting `data-live-action-param` / `data-live-prop` with untrusted values — bound back to the server unchecked.
-            Do NOT flag:
-            - Default `{{ value }}` interpolation — Twig auto-escapes for the active context.
-            - `|raw` on values originating from a `Markdown` / `Sanitize` (HtmlSanitizer) transformation upstream — trust the sanitizer unless evidence shows otherwise.
-            - `{% component %}` / `<twig:…/>` with statically-typed props in a `LiveComponent` whose writable props are constrained by validators.
-            </skills>
-            SKILL,
-        ProjectFileType::CONFIG->value => <<<'SKILL'
-            <skills role="config">
-            Hunt:
-            - Hardcoded secrets, API keys, DSNs (look for `_KEY`, `_SECRET`, `_TOKEN`, `password:`, full DSN strings).
-            - `security.firewalls` with `security: false` on protected paths, or `access_control` rules with overly broad path patterns.
-            - `framework.session.cookie_secure: false` / `cookie_samesite: none` / missing `cookie_httponly` in non-test envs.
-            - Dev-only routes (`_profiler`, `_wdt`) not gated by environment or IP allowlist.
-            - `cors.allow_origin: '*'` combined with `allow_credentials: true` — credential leak.
-            - Exposed services with public `true` that wrap dangerous primitives (filesystem, process, raw SQL).
-            - `framework.rate_limiter` absent for login / password-reset / API endpoints declared elsewhere.
-            - `framework.messenger.transports.*.serializer: 'php_serialize'` or any class implementing `SerializerInterface` that calls `unserialize` on dequeued payloads.
-            - `framework.webhook` declared without an HMAC secret env var or with the secret committed in plaintext.
-            - `framework.lock` not configured for routes that perform balance-affecting operations.
-            - `framework.html_sanitizer` set with `allowAllStaticAttributes()` or wide `allowElement()` lists — XSS via "sanitized" output.
-            - `framework.http_client.scoped_clients.*.verify_peer: false` / `verify_host: false` — MITM exposure.
-            - `messenger.routing` directing privileged commands to a transport with `failure_transport: null` — silent loss of audit trail.
-            Do NOT flag:
-            - Environment-variable references like `%env(DATABASE_URL)%` — those are externalized, not hardcoded.
-            - `_profiler` / `_wdt` declarations gated by `when@dev` / `when@test`.
-            - `framework.messenger.transports.*.serializer: messenger.transport.symfony_serializer` — the safe default.
-            </skills>
-            SKILL,
-        ProjectFileType::PHP->value => <<<'SKILL'
-            <skills role="php">
-            Hunt:
-            - Symfony ExpressionLanguage / Security Expression evaluating strings derived from user input.
-            - `HttpClientInterface` calls with user-controlled URL or host — SSRF; check for protocol & host allowlist; `max_redirects` not bounded + redirect host not re-validated.
-            - `unserialize()` / `igbinary_unserialize()` on untrusted payloads (cache values, queue messages, request body).
-            - `Process` / `proc_open` / `shell_exec` constructed with user input concatenation; no `escapeshellarg`.
-            - PSR-3 logger sinks receiving raw request payloads without redaction (log injection, log forging).
-            - Cryptography: `md5`/`sha1` for security purposes, `random_int` vs `rand`/`mt_rand`, hardcoded IVs/keys.
-            - `MailerInterface::send()` with `Email::from($userInput)` / `subject($userInput)` / `addBcc($userInput)` — header injection via newline in user data.
-            - `CacheInterface::get($key, ...)` where `$key` is derived from request input without normalization — cache poisoning / cross-tenant leak.
-            - `CacheItemPoolInterface` reads writing user-derived payloads then trusting them on subsequent reads (poisoned cache → privilege bypass).
-            - `LockFactory::createLock()` missing on critical sections (refund, balance mutation, idempotent webhook) — race condition.
-            - `HtmlSanitizerInterface::sanitize()` configured with `allowElement('script')` / `allowAttribute('on*')` — sanitizer disarmed.
-            - `RateLimiterFactory::create($key)` with `$key` constant (`'global'`) instead of per-user/per-IP — limiter trivially exhausted.
-            - PSR-16 / PSR-6 store handing back unserialized objects when the cache backend was filled with user input.
-            Do NOT flag:
-            - `md5`/`sha1` on non-security data (cache keys, ETags, file fingerprints) — those are integrity, not authentication.
-            - `Process` invocations with hardcoded argument arrays (`new Process(['ls', '-la'])`) — no shell interpolation occurs.
-            - `HttpClient::request()` against `%env(INTERNAL_SERVICE_URL)%` — that is an externally-configured trusted host, not SSRF.
-            </skills>
-            SKILL,
-    ];
 
     /**
      * @param list<ProjectFile> $files
@@ -644,13 +357,6 @@ final readonly class AttackerPromptBuilder implements AttackerPromptBuilderInter
             $files,
         );
 
-        $blocks = [];
-        foreach (self::SKILL_PRIORITY as $type) {
-            if ($this->emitAllSkills || \in_array($type, $presentTypes, true)) {
-                $blocks[] = self::SKILLS[$type->value];
-            }
-        }
-
-        return implode("\n\n", $blocks);
+        return $this->attackerSkillRegistry->render($presentTypes, $this->emitAllSkills);
     }
 }
